Report why a site failed in report:users, and whose failure it is (#9810)
A failing site printed only its exit code, and printed it as a verdict:
`✖ [33/87] hshp.uiowa.edu (exit 1)`. Both halves misled. The exit code
alone doesn't distinguish a broken site from an unreachable one, and the
line described one attempt as though it were the outcome — the pool
retries failures and reports progress per attempt, so a site that failed
once and succeeded on retry still made it into the report, contradicting
its own ✖.

- Move failureReason() out of multisite:execute into a
  DescribesDrushFailures trait both fleet commands use. It already had
  the message-first rendering report:users wanted; keeping a second copy
  is what let them drift apart in the first place.
- report:users leads with drush's message in the progress line and the
  end-of-run summary, and the progress line now says "attempt failed" —
  the summary, built from final results, is the verdict.
